env: refactoring
