+++ component/profile/profile.tsx
@@ -5,6 +5,7 @@ interface ProfileProps {
     avatarSize?: string;
     nameSize?: string
     padding?: string;
+    date?: string;
     trailing?: any;
 }
 
@@ -16,9 +17,15 @@ export default function Profile(props: ProfileProps) {
                         "258867831_481563193336979_1631726358864355454_n.jpg?stp=dst-jpg_p480x480&_nc_cat=102&ccb=1-7" +
                         "&_nc_sid=c6021c&_nc_ohc=RmAVRG8XacgAX94_ZNV&_nc_ht" +
                         "=scontent-nrt1-1.xx&oh=00_AT-UhiZrDAfSfOTHhN0-dJ5w4T-wHaxdU40zvog5GXQ0Fg&oe=6312F3EE"}/>
-            <Text>
-                이름
-            </Text>
+
+            <Column>
+                <Text>
+                    이름
+                </Text>
+                {
+                    props.date != null ? <Date>{props.date}</Date> : null
+                }
+            </Column>
             {props.trailing}
         </Container>
     );
@@ -28,8 +35,16 @@ interface ContainerProps {
     padding?: string;
 }
 
+const Column = styled.span`
+  display: flex;
+  flex-direction: column;
+`;
+
 const Container = styled.div<ContainerProps>`
   padding: ${(props) => props.padding};
+  display: flex;
+  align-items: center;
+  flex-direction: row;
 `;
 
 interface TextProps {
@@ -40,4 +55,11 @@ interface TextProps {
 const Text = styled.text<TextProps>`
   margin: 0 8px;
   font-size: ${(props) => props.fontSize};
+`;
+
+const Date = styled.text`
+  margin: 0 8px;
+  font-size: 12px;
+  font-weight: lighter;
+  color: #A7A7A7;
 `;
