+++ stories/component/header.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {ComponentStory, ComponentMeta} from '@storybook/react';
-import Header from "../component/header/header";
+import Header from "../../component/header/header";
 
 export default {
     title: 'codepie/Header',
